+++ alacritty/src/display/window.rs
@@ -41,8 +41,8 @@ use winit::window::{
 use alacritty_terminal::index::Point;
 
 use crate::cli::WindowOptions;
-use crate::config::window::{Decorations, Identity, WindowConfig};
 use crate::config::UiConfig;
+use crate::config::window::{Decorations, Identity, WindowConfig};
 use crate::display::SizeInfo;
 
 /// Window icon for `_NET_WM_ICON` property.
@@ -198,7 +198,7 @@ impl Window {
         use_srgb_color_space(&window);
 
         let scale_factor = window.scale_factor();
-        log::info!("Window scale factor: {}", scale_factor);
+        log::info!("Window scale factor: {scale_factor}");
         let is_x11 = matches!(window.window_handle().unwrap().as_raw(), RawWindowHandle::Xlib(_));
 
         Ok(Self {