+++ alacritty_terminal/src/index.rs
@@ -1,7 +1,7 @@
 //! Line and Column newtypes for strongly typed tty/grid/terminal APIs.
 
 /// Indexing types and implementations for Grid and Line.
-use std::cmp::{max, min, Ord, Ordering};
+use std::cmp::{Ord, Ordering, max, min};
 use std::fmt;
 use std::ops::{Add, AddAssign, Deref, Sub, SubAssign};