+++ alacritty_terminal/src/tty/mod.rs
@@ -33,6 +33,13 @@ pub struct Options {
 
     /// Extra environment variables.
     pub env: HashMap<String, String>,
+
+    /// Specifies whether the Windows shell arguments should be escaped.
+    ///
+    /// - When `true`: Arguments will be escaped according to the standard C runtime rules.
+    /// - When `false`: Arguments will be passed raw without additional escaping.
+    #[cfg(target_os = "windows")]
+    pub escape_args: bool,
 }
 
 /// Shell options.
@@ -94,10 +101,10 @@ pub fn setup_env() {
     // default to 'xterm-256color'. May be overridden by user's config
     // below.
     let terminfo = if terminfo_exists("alacritty") { "alacritty" } else { "xterm-256color" };
-    env::set_var("TERM", terminfo);
+    unsafe { env::set_var("TERM", terminfo) };
 
     // Advertise 24-bit color support.
-    env::set_var("COLORTERM", "truecolor");
+    unsafe { env::set_var("COLORTERM", "truecolor") };
 }
 
 /// Check if a terminfo entry exists on the system.